% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{alluvialplot}
\alias{alluvialplot}
\title{Create an alluvial plot from long categorical data}
\usage{
alluvialplot(labelsdf, fromlevels, tolevels, ggfittext)
}
\arguments{
\item{labelsdf}{data.frame with two columns of raw categorical label: for example, each row is a cell (or other observation), and each column is metadata column 1 and metadata column 2}

\item{fromlevels}{character vector. levels of labelsdf first column. can be used to set a desired order from first (top) to last (bottom)}

\item{tolevels}{character vector. levels of labelsdf second column. can be used to set a desired order from first (top) to last (bottom)}

\item{ggfittext}{T/F - whether to use ggfittext, to try to squeeze or remove tiny stratum labels}
}
\value{
a ggplot object
}
\description{
Wrapper around ggalluvium package for ggplot based alluvial plot, for quickly making alluvial plot from "long", "raw" categorical data (such as Seurat object meta.data), rather than two-way counts of categories.
}
\examples{
#' labelsdf can look like this:
                     From      To
AACCCAAGCATGCGA-1    Malignant  2
AAACCCAAGTAGGTTA-1    Malignant  2
AAACCCACAAAGCACG-1   Neutrophil  0
AAACCCACACACTTAG-1 FILTERED_OUT  3
AAACCCACAGCAGTAG-1    Malignant  2
AAACCCACATACCGTA-1    Malignant  2
}
