% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{doubletfinderwrapper}
\alias{doubletfinderwrapper}
\title{DoubletFinder Wrapper}
\usage{
doubletfinderwrapper(seuratobject, clusters)
}
\arguments{
\item{seuratobject}{A Seurat object.}

\item{clusters}{string, should match a column name of seuratobject metadata with clusters or other cell group annotations. default = 'seurat_clusters'}
}
\value{
a data.frame with barcodes and DoubletFinder classifications
}
\description{
Models homotypic doublets using the following table:
https://kb.10xgenomics.com/hc/en-us/articles/360001378811-What-is-the-maximum-number-of-cells-that-can-be-profiled-
}
